backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-Mail==0.9.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
APScheduler==3.10.4
python-dotenv==1.0.0
bcrypt==4.1.2
eventlet==0.35.1
